backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.3.3
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.0.0